:fire: merge image_lib to vision_lib
